Fix client API version.
The GBP client CLI was failing, with the following error message:
"Invalid network client version '{'network': '2.0'}'. must be one of: 2.0".
A change in the openstack SDK (>=18.0) was causing this failure in the
upstream neutron python client, which was resolved with this commit:
https://review.opendev.org/#/c/615184/

This commit makes a similar fix for the GBP python client.

Change-Id: I660c6881ef406a6d9995e2712399a3ac06ba207a