Make gate jobs non-voting
Rocky is long since EOL, and finally no longer works in the gate.
This patch makes the gate jobs non-voting, This is low-risk, as
the queens job still works, as does stein, and rocky should be
a clean backport. This is only needed until these branches finally
go fully EOL for our project.

Change-Id: Ic383395a776db8415af690baf4c4f0a9825beacc